Better build config
Puts built version in a similar folder structure to the app server. This allows for better testing of the built app including the addition of a `grunt serve-built` command.

Reconfigures `baseUrl` for built versions to work with the weird folder structure mentioned above.

Don't copy `.styl` files to the built version of the app. I was running into a problem with my simple web server auto-compiling them and overwriting `App.css`.

Fixes #22